Updated Contact Form type tel move into component and store image in public folder
